Partially revert "Add initial tests for update_{llc_,cc_,}test_checks.py"
This reverts part of commit ead9874ef2b5126f8ac3edfcdab83ec0599f4aaa.
It reverts 08208777bdb9a9d63348e86923b60340e656e056.

We currently run LLVM tests in environments where python3 exists on
PATH, but it is broken. I don't think PATH discovery is a strong enough
signal that a working Python 3 installation exists.

If this will be the way forward, IMO we should follow the direction of
debug-info-tests, and use CMake's PYTHON_EXECUTABLE, which in the near
future will be a known-to-work Python 3 executable. If it's not Python
3, then we don't have to run this test.